refactor(ui): TableListを移行
