+++ packages/dash/src/lib/common/omuchat/client.ts
@@ -1,5 +1,6 @@
 import type { Chat } from '@omuchatjs/chat';
 import type { Client as OmuClient } from '@omuchatjs/omu';
+import { client } from '@omuchatjs/ui';
 import { getContext, setContext } from 'svelte';
 import type { Dashboard } from './dashboard.js';
 
@@ -12,6 +13,7 @@ interface ClientContext {
 }
 
 export function setClient(context: ClientContext) {
+    client.set(context.client);
     return setContext<ClientContext>(CLIENT_CONTEXT, context);
 }
