+++ packages/dash/src/lib/main/page/assets/PageAssets.svelte
@@ -6,7 +6,7 @@
     import InputTextLazy from '$lib/common/input/InputTextLazy.svelte';
     import type { Asset } from '$lib/common/omuchat/asset.js';
     import { getClient } from '$lib/common/omuchat/client.js';
-    import TableList from '$lib/common/omuchat/TableList.svelte';
+    import { TableList } from '@omuchatjs/ui';
 
     const { dashboard } = getClient();
