+++ packages/dash/src/lib/main/page/home/PanelApps.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
     import AppEntry from './AppEntry.svelte';
 
-    import TableList from '$lib/common/omuchat/TableList.svelte';
     import { getClient } from '$lib/common/omuchat/client.js';
     import { tauriWindow } from '$lib/utils/tauri.js';
     import { onMount } from 'svelte';
 
     import type { AppMetadata } from '$lib/common/omuchat/app-metadata.js';
     import { t } from '$lib/i18n/i18n-context.js';
+    import { TableList } from '@omuchatjs/ui';
 
     export let filter: (key: string, app: AppMetadata) => boolean = () => true;
     const { dashboard } = getClient();
