+++ packages/dash/src/lib/main/panel/channels/PanelChannels.svelte
@@ -9,10 +9,9 @@
 
     import ButtonMini from '$lib/common/input/ButtonMini.svelte';
     import { getClient } from '$lib/common/omuchat/client.js';
-    import TableList from '$lib/common/omuchat/TableList.svelte';
     import { screenContext } from '$lib/common/screen/screen.js';
     import ScreenSetup from '$lib/main/setup/ScreenSetup.svelte';
-    import { Tooltip } from '@omuchatjs/ui';
+    import { TableList, Tooltip } from '@omuchatjs/ui';
 
     export let filter: (key: string, message: models.Channel) => boolean = () => true;
