+++ packages/dash/src/lib/main/panel/rooms/PanelRooms.svelte
@@ -8,9 +8,9 @@
 
     import Button from '$lib/common/input/Button.svelte';
     import { getClient } from '$lib/common/omuchat/client.js';
-    import TableList from '$lib/common/omuchat/TableList.svelte';
     import { screenContext } from '$lib/common/screen/screen.js';
     import ScreenSetup from '$lib/main/setup/ScreenSetup.svelte';
+    import { TableList } from '@omuchatjs/ui';
 
     export let filter: (key: string, room: models.Room) => boolean = () => true;
     export let sort: (a: models.Room, b: models.Room) => number = (a, b) => {
