+++ packages/ui/src/lib/stores.ts
@@ -1,3 +1,4 @@
+import type { Client } from "@omuchatjs/omu";
 import { writable, type Writable } from "svelte/store";
 
 type TranslateFunction = (key: string, options?: Record<string, unknown>) => string;
@@ -6,4 +7,4 @@ export const translate: Writable<TranslateFunction> = writable((key: string, opt
     return `(translation for ${key} not found. options: ${JSON.stringify(options)})`;
 });
 
-
+export const client: Writable<Client> = writable();
